requirements.txt
wordcloud==1.9.2
jieba==0.42.1
numpy==1.24.4
Pillow==10.1.0